src/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Login==0.6.3
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
mysqlclient==2.2.0
packaging==23.2
protobuf==4.21.12
PyMySQL==1.1.0
pyodbc==5.0.1
Werkzeug==3.0.1
WTForms==3.1.1
